+++ docs/CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Added
+-----
+- Add `test_data_output` management command to test that files referenced in
+  data output exist in the database
+
+
 ===================
 42.1.0 - 2024-12-16
 ===================